chgt decimal/int col height

+++ src/users/entities/user.entity.ts
@@ -52,7 +52,7 @@ export class Users {
 
   @Column({
     nullable: false,
-    type: 'int',
+    type: 'decimal',
     width: 6,
   })
   height: number;
